suite

+++ backend/controllers/post.controller.js
@@ -7,6 +7,7 @@ module.exports.getPosts = async (req, res) => {
     try {
         const posts = await postModel.find().populate('author', 'username').populate('likers', 'username');
         res.status(200).json(posts);
+
     } catch(error) {
         res.status(500).json({ message: 'Erreur lors de la récupération des messages', error: error.message});
     }
@@ -27,6 +28,7 @@ module.exports.setPosts = async (req, res) => {
         })
 
         res.status(201).json(post);
+
     } catch(error) {
         res.status(500).json({ message: 'Erreur lors de la création du message.', error: error.message });
     }
@@ -44,8 +46,6 @@ module.exports.editPost = async (req, res) => {
         const userIdObject = new mongoose.Types.ObjectId(req.userId);
 
         // Vérifier si l'utilisateur connecté est bien l'auteur du post
-        console.log('post.author:', post.author);
-        console.log('req.userId:', req.userId);
 
         if (!post.author.equals(userIdObject)) {
             return res.status(403).json({ message: "Accès refusé. Vous n'êtes pas l'auteur de ce message." });
@@ -58,6 +58,7 @@ module.exports.editPost = async (req, res) => {
         );
 
         res.status(200).json(updatedPost);
+
     } catch (error) {
         res.status(500).json({ message: 'Erreur lors de la mise à jour du message.', error: error.message });
     }
@@ -72,8 +73,15 @@ module.exports.deletePost = async (req, res) => {
             return res.status(404).json({ message: "Ce post n'existe pas." });
         }
 
+        const userIdObject = new mongoose.Types.ObjectId(req.userId);
+
+        if (!post.author.equals(userIdObject)) {
+            return res.status(403).json({ message: "Accès refusé. Vous n'êtes pas l'auteur de ce message." });
+        }
+
         await postModel.findByIdAndDelete(req.params.id);
-        res.status(200).json({ message: "Message supprimé " + req.params.id });
+        res.status(200).json({ message: "Message supprimé " });
+
     } catch (error) {
         res.status(500).json({ message: 'Erreur lors de la suppression du message.', error: error.message });
     }
@@ -82,17 +90,27 @@ module.exports.deletePost = async (req, res) => {
 // liker un message
 module.exports.likePost = async (req, res) => {
     try {
-        const post = await postModel.findByIdAndUpdate(
-            req.params.id,
-            { $addToSet: { likers: req.body.userId } },
-            { new: true }
-        ).populate('author', 'username').populate('likers', 'username');
-
-        if (!post) {
-            return res.status(404).json({ message: "Ce post n'existe pas." });
+        const userId = req.userId;
+        const post = await postModel.findById(req.params.id);
+        // travailler cette partie
+        if (!userId) {
+            return res.status(400).json({ message: "utilisateur non trouvé" });
         }
 
+        if (!post) {
+            return res.status(404).json({ message: "post non trouvé"});
+        }
+
+        if (post.likers.includes(userId)) {
+            return res.status(400).json({ message: "post déjà liké"});
+        }
+
+        // ajouter l'utilisateur au tableau des likes
+        post.likers.push(userId);
+        await post.save();
+
         res.status(200).json(post);
+
     } catch (error) {
         res.status(500).json({ message: 'Erreur lors du like du message.', error: error.message });
     }
@@ -101,17 +119,23 @@ module.exports.likePost = async (req, res) => {
 // disliker un message
 module.exports.dislikePost = async (req, res) => {
     try {
-        const post = await postModel.findByIdAndUpdate(
-            req.params.id,
-            { $pull: { likers: req.body.userId } }, 
-            { new: true }
-        ).populate('author', 'username').populate('likers', 'username');
+        const userId = req.userId;
+        const post = await postModel.findById(req.params.id);
 
         if (!post) {
             return res.status(404).json({ message: "Ce post n'existe pas." });
         }
 
+        if(!post.likers.includes(userId)) {
+            return res.status(400).json({ message: "vous n'avez pas liké ce post, vous ne pouvez pas le dislike."})
+        }
+
+        // enlever l'utilisateur du tableau des likes
+        post.likers = post.likers.filter(likerId => likerId.toString() !== userId);
+        await post.save();
+
         res.status(200).json(post);
+
     } catch (error) {
         res.status(500).json({ message: 'Erreur lors du dislike du message.', error: error.message });
     }
